requirements.txt
# Core Dependencies
python-dotenv
httpx

# Module 1: NiruSpider (Crawler)
scrapy
requests
beautifulsoup4
feedparser  # RSS feed parsing
lxml
celery

# Module 2: NiruParser (ETL & Embedding)
trafilatura  # HTML text extraction
newspaper3k  # Alternative article extraction
PyPDF2
pdfplumber
pdf2image  # PDF to image conversion for Vision RAG
langchain
langchain-community
sentence-transformers>=2.2.0
huggingface-hub>=0.20.0
tiktoken  # Token counting
textblob  # Sentiment analysis
youtube-transcript-api  # YouTube transcript extraction
cohere  # Cohere Embed-4 for multimodal embeddings

# Module 3: NiruDB (Vector Store)
chromadb
faiss-cpu  # Alternative: faiss-gpu for GPU support
loguru
elasticsearch  # For document storage and search
upstash-redis  # For cache and vector storage
upstash-vector  # For vector database
qdrant-client  # Fallback vector store
cryptography  # For encrypting API keys
redis  # For rate limiting and caching

# Database (PostgreSQL for sessions, feedback, chat history)
psycopg2-binary
sqlalchemy
alembic  # Database migrations

# Module 4: NiruAPI (RAG Interface)
fastapi
uvicorn[standard]
pydantic
python-multipart
africastalking  # SMS gateway for Kenya

# Module 8: NiruAuth (Authentication & Authorization)
PyJWT>=2.8.0  # JWT token handling
bcrypt>=4.0.0  # Password hashing
email-validator>=2.0.0  # Email validation for Pydantic

# LLM Integration
openai # For OpenAI models and Moonshot AI (compatible API)
anthropic  # For Claude
google-generativeai  # For Gemini AI
# huggingface-hub is already specified above

# Utilities
pyyaml
pandas
numpy
tqdm
schedule  # For automated crawling
celery  # Async task queue
colorama  # Colored terminal output
loguru  # Advanced logging
psutil  # System monitoring
reportlab
python-docx

# Module 6: NiruVoice (Real-time Communication)
livekit-agents[openai,assemblyai,silero]
livekit
livekit-plugins-noise-cancellation
tenacity>=8.2.0  # Retry logic with exponential backoff
circuitbreaker>=1.4.0  # Circuit breaker pattern
prometheus-client>=0.19.0  # Prometheus metrics (optional)
deepgram-sdk>=3.0.0  # Deepgram STT
elevenlabs>=0.2.0  # ElevenLabs TTS

# Module 7: NiruHybrid - Hybrid Conv-Transformer Pipeline
torch # PyTorch for neural networks
transformers>=4.30.0  # Hugging Face transformers (compatible with huggingface-hub>=0.20.0)
bitsandbytes  # Quantization support
accelerate  # Mixed precision training
diffusers  # Diffusion model framework
einops # Tensor operations
streamlit  # Optional: for monitoring dashboard

# Agentic AI & LangGraph
langgraph>=0.2.0
langchain-core>=0.3.0

# External Data Sources
twikit>=0.1.0
google-search-results>=2.4.0
ddgs
rank-bm25>=0.2.2

# Development
pytest
black
flake8
